+++ projects/amdsmi/rust-interface/examples/amdsmi_exporter/http_server.rs
@@ -21,17 +21,17 @@
 use crate::amdsmi_collectors::AmdsmiCollectors;
 use axum::response::{IntoResponse, Response};
 use axum::{routing::get, Router};
-use hyper::Server;
 use prometheus_client::encoding::text::encode;
 use prometheus_client::registry::Registry;
 use std::net::SocketAddr;
 use std::sync::Arc;
+use tokio::net::TcpListener;
 use tokio::sync::Mutex;
 
 async fn serve_req(registry: Arc<Mutex<Registry>>) -> impl IntoResponse {
     let mut buffer = String::new();
     let registry = registry.lock().await;
-    encode(&mut buffer, &*registry).unwrap();
+    encode(&mut buffer, &registry).unwrap();
     Response::builder()
         .header("Content-Type", "text/plain; version=0.0.4")
         .body(buffer)
@@ -39,10 +39,10 @@ async fn serve_req(registry: Arc<Mutex<Registry>>) -> impl IntoResponse {
 }
 
 pub async fn run_http_server(collectors: &Arc<Mutex<AmdsmiCollectors>>, addr: SocketAddr) {
-    let app = Router::new().route(
+    let app: Router = Router::new().route(
         "/metrics",
         get({
-            let collectors = Arc::clone(&collectors);
+            let collectors = Arc::clone(collectors);
             move || {
                 let collectors = Arc::clone(&collectors);
                 async move {
@@ -56,8 +56,10 @@ pub async fn run_http_server(collectors: &Arc<Mutex<AmdsmiCollectors>>, addr: So
 
     println!("Listening on http://{}", addr);
 
-    Server::bind(&addr)
-        .serve(app.into_make_service())
+    let listener = TcpListener::bind(addr)
         .await
-        .unwrap();
+        .expect("failed to bind TCP listener");
+    axum::serve(listener, app)
+        .await
+        .expect("axum server failed");
 }